# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank Project API Documentation",
        "description": "API endpoints for managing departments, products, and product types in the Bank Project.",
        "contact": {
            "email": "tvoj.email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Bank Project API Server"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/product_types": {
            "get": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Get all product types",
                "description": "Returns a list of all product types.",
                "operationId": "4c15fcb7e4cdc99700c94d79716b12fa",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup - potrebna Basic HTTP autentifikacija."
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "post": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Create a new product type",
                "description": "Creates a new product type record.",
                "operationId": "f3c6a1e72cd0fae8c796a41ed6246af4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "PRODUCT_TYPE_CD"
                                ],
                                "properties": {
                                    "PRODUCT_TYPE_CD": {
                                        "type": "string",
                                        "example": "LOAN"
                                    },
                                    "NAME": {
                                        "type": "string",
                                        "example": "Loan Product Type",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product Type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "PRODUCT_TYPE_CD": [
                                                    "The PRODUCT_TYPE_CD has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            }
        },
        "/api/product_types/{product_type_cd}": {
            "get": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Get product type by code",
                "description": "Returns a single product type by its code.",
                "operationId": "025e0febbdc21b4a26a7990579de0830",
                "parameters": [
                    {
                        "name": "product_type_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_TYPE_CD) of the product type to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product Type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "put": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Update an existing product type",
                "description": "Updates an existing product type record by code.",
                "operationId": "7cee6303c80d52a25fdebcaec14fd77a",
                "parameters": [
                    {
                        "name": "product_type_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_TYPE_CD) of the product type to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "NAME": {
                                        "type": "string",
                                        "example": "New Loan Product Type Name",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product Type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": {
                                                    "NAME": [
                                                        "The NAME field must be a string."
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "delete": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Delete a product type",
                "description": "Deletes a product type record by code.",
                "operationId": "4bc077ecb7c5f448f0b5c1fe7f006758",
                "parameters": [
                    {
                        "name": "product_type_cd",
                        "in": "path",
                        "description": "Code (PRODUCT_TYPE_CD) of the product type to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product Type deleted successfully (No Content)"
                    },
                    "404": {
                        "description": "Product Type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductType": {
                "title": "ProductType",
                "description": "Product Type model",
                "properties": {
                    "PRODUCT_TYPE_CD": {
                        "title": "Product Type Code",
                        "description": "Unique code for the product type",
                        "type": "string",
                        "example": "ACC"
                    },
                    "NAME": {
                        "title": "Name",
                        "description": "Name of the product type",
                        "type": "string",
                        "example": "Account"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductType"
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "HTTP Basic Authentication (Koristi e-mail i lozinku testnog korisnika)"
            }
        }
    },
...
    "security": [
        {
            "basicAuth": []
        }
    ]
}